Add a generic regex parser
This adds a new parser that allows users to configure any number of
regexes to parse custom formats. These are stored in a `.yaml` file
which can be configured using `--regexes path/to/file.yaml` or by
setting `$WD_REGEXES`.

Fields are extracted using named capture groups, and aside from the
3 common fields (`datetime`, `level`, `text`), other named capture
groups are added to the metadata map.

Includes lots of additional plumbing to make config available to
parsers.