Add optional support for message reordering
This adds a new flag, `--ordered`, to attempt to reorder messages
before writing them out. Incoming messages are stored in a min-heap
and ordered by their semantic timestamps (either embedded in the
format or provided by a reader), held for a configurable period of
time (1 second by default, configurable with `--buffer-ms`), and
re-emitted in order as the deadline expires.

While this isn't a perfect solution to out-of-order messages, it
should significantly improve output with datasources that have high
quality timestamps, like the Kubernetes reader.

Note that messages that arrive too late (outside the `--buffer-ms`
window) may still be written out of order; however, they'll at least
be written as close to their ideal position as possible. Increasing
the buffer can prevent this at the cost of higher latency.

Signed-off-by: Tim Buckley <timothyb89@gmail.com>